Issue 14: GET:made changes
