+++ package.json
@@ -33,6 +33,7 @@
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.6.1",
     "jest": "^22.1.4",
+    "jest-json-schema": "^1.2.0",
     "sequelize-cli": "^3.2.0",
     "supertest": "^3.0.0"
   },
