+++ src/routes/index.js
@@ -1,0 +1,3 @@
+const banks = require('./api/max-amount.js');
+
+module.exports = [].concat(banks);
